Add close button to lineage tab (#117)
Add a close button to the toolbar in the lineage tab. Use redux state to
track the open / close state of the tab as the state is now needed
globally. Move tooltips in the toolbar to the top to not cover other
items. Fix / refactor tooltip of graphViewSelector item to not show two
tooltips.